Fix non-SIMD (MSan) builds
Previously, WITH_SIMD was always defined and we were simply switching
the implementation to jsimd_none.c for MSan.

The placeholder jsimd_none.c was removed in [1], so now we have to
control SIMD via defines.

This CL removes the hard-coded WITH_SIMD defines and consolidates all
SIMD configuration in the "simd" library.

[1] https://github.com/libjpeg-turbo/libjpeg-turbo/commit/e8b40f3c2ba187ba95c13c3e8ce21c8534256df7

Bug: 398911117
Change-Id: Ia8108469c312a1b9b5c9db1868556214e6db7599