libjpeg-turbo: base/hash/md5 -> crypto/obsolete/md5
Migrates from the deprecated base/hash/md5 API to the non-deprecated one
in crypto/.

We're using MD5 to be consistent with the known-good values provided by
upstream. If we move to some other hash function it becomes harder to
update if the implementation (and the known-good hashes) change
upstream, which has happened in the past.

There should be no change in behavior.

Bug: 406729261
Change-Id: I666d011a88d461295d91f30e265367df87fa4ab9